library: fix passphrase_temporary idempotency
When passphrase_temporary is set, we remove the specific passphrase
given as argument -- which can be provided via either passphrase or
key_file parameters.

If we run the role the 2nd time with the same arguments, it was
reporting an error, since the provided passphrase is not valid anymore.
This is true, but in this case, we now try to retrieve a valid
passphrase from existing bindings, and error out only in cases where
we are unable to do so.