Add dracut module for disabling autoconnect within initrd
The network connection created during the initramfs for unlocking the
device may conflict with the actual system connection later on, when
the system attempts to configure the network as per the regular
configuration, after the boot process has completed.

To prevent this situation, we now deploy a dracut module to disable
autoconnect of the network connectons that are active during the
initramfs.