network-flush: reset autoconnect-priority to zero (#65)
With older dracut, where NetworkManager does not run as a daemon
and with dbus, the dracut module we have will not work to disable
autoconnect for the connections created within the initramfs, which
may be a problem as after the network flush, NM will activate the
connection with the higher priority, and the initramfs connection
may be it.

Some connections are created with negative priority, by e.g. the
Anaconda installer, which may be a problem in this situation. To
work around this issue, we now reset the autoconnect priority of
the connections to zero.