Do not report password in stacktrace or return value from module (#98)
Obscure the encryption_password when returning the results to the caller
of the module.  If the user needs the password, they can get it from the
original arguments.

Catch exceptions so that the password can be obscured from the return
value.

Create a test to cause a NbdeClientClevisError so that we can verify
the error message is correct and the password is obscured.

Use ansible vault for testing the password.