Fix nbde_client error handling (#101)
Because the nbde_client role throws the error in a block/always, the error
must be caught with rescue and rethrown.  See
https://richm.github.io/how-to-catch-and-reraise-errors-in-ansible

The sensitive field might be in a deeply nested data structure, so
iterate through the data structure.

The re-raised error is not in `ansible_failed_result` it is in
`ansible_failed_result.msg`

The format of `ansible_failed_result` and the error message is
different on EL7 than it is on other platforms - handle the
difference.