fix: idempotence issue when binding fails to be added
Sometimes, the role will not be able to add the required bindings, in
which case it is expected to rollback and undo any change it has done.

In certain cases, the rollback was not performed correctly, and caused
idempotence issues. We fix that by performing a backup of the LUKS
header before doing the operations, so that we can properly restore it
in the case the operation cannot be completed successfully.

Signed-off-by: Sergio Correia <scorreia@redhat.com>